+++ lib/ui/auth/view/auth_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../../bbs/view/bbs_page.dart';
-import '../../common/error_page.dart';
+import 'error_page.dart';
 import '../view_model/auth_view_model.dart';
 import 'sign_in_page.dart';
